move otu commands to submodule
